lisp: kernelize the runtime as SysLisp (K4 — boots in the OS)
Wrap the host-proven Scheme runtime (K1-K3) as a signed Sys* module that
runs inside Cardinal; and self-tests the whole stack on real boot.

- modules/SysLisp: links the `lisp` static lib; module_init does
  lisp_gc_init(__builtin_frame_address(0)) as the conservative-GC stack
  base, routes the output sink to the COM1 debug log (print_str, chunked),
  exposes a foreign-function (uptime-ns) calling the kernel timer, and runs
  an in-OS self-test. All 11 cases pass on boot: arithmetic/recursion/map/
  tail loops, flonums (3.75, 3.5 -- SSE in task context), display->serial,
  GC survival across forced collections, the FFI call, and the cooperative
  scheduler + copy-on-send + per-context-GC producer/consumer (-> 15).
- It loads from servicescript, not loadscript: the boot thread never returns
  past SysTaskMgr's scheduler handoff, and the runtime's hardware doubles are
  only legal in task context (servicescript runs as a task). Loaded after
  SysDebug/SysTimer/SysFP/SysMemory so print_str/timer/FP/malloc are up.
- Bare-metal SSE fixes: the lisp lib builds with -mstackrealign (the -mno-sse
  kernel call path doesn't keep the 16-byte stack alignment SSE assumes) and
  -fno-vectorize/-fno-slp-vectorize; print.c's emit_flonum keeps the integer
  part a signed int64 so the (double)ip back-conversion is a cvtsi2sd instead
  of the uint64->double magic-constant sequence (movapd of a 16-byte .rodata
  constant), whose aligned load #GP'd because the module loader does not
  16-align section data.
- sched.c: lisp_sched_run now clears g_sched on return so it never dangles
  past a run (review).

Deferred to K5: the native-ISR -> event -> wake-context bridge (needs the
scheduler to be the persistent per-core loop; the recv/wake software path it
rides on is already proven in K2). Host suite + freestanding stay green.

+++ libs/lisp/src/print.c
@@ -68,12 +68,15 @@ static void emit_flonum(sink *s, double v) {
     }
     if (neg)
         emit_ch(s, '-');
-    // Normalize very large magnitudes with a decimal exponent so the integer-
-    // part cast below cannot overflow uint64 (that conversion would be UB). The
-    // exponent is appended as a trailing eN. This path is intentionally coarse;
-    // a shortest-round-trip printer is a later refinement.
+    // Normalize very large magnitudes with a decimal exponent so the integer-part
+    // cast below stays within int64 range (that conversion would otherwise be UB).
+    // ip is kept SIGNED (int64) on purpose: the (double)ip back-conversion below
+    // is then a single cvtsi2sd. An unsigned (uint64)->double would instead lower
+    // to the SSE2 magic-constant sequence (movapd of a 16-byte .rodata constant +
+    // subpd), whose ALIGNED load #GPs because the kernel module loader does not
+    // 16-align section data. The exponent is appended as a trailing eN.
     int e10 = 0;
-    while (v >= 9.223372036854775808e18) {  // 2^63: keep ip in int64 range for emit_int
+    while (v >= 9.223372036854775808e18) {  // 2^63: keep ip in int64 range
         v /= 10.0;
         e10++;
     }
@@ -82,8 +85,8 @@ static void emit_flonum(sink *s, double v) {
     for (int k = 0; k < FRAC; k++)
         rounding /= 10.0;
     v += rounding;  // round at the FRAC-th fractional digit (may carry into ip)
-    uint64_t ip = (uint64_t)v;
-    emit_int(s, (int64_t)ip);
+    int64_t ip = (int64_t)v;
+    emit_int(s, ip);
     emit_ch(s, '.');
     double frac = v - (double)ip;
     char digits[16];
